+++ src/sflow.c
@@ -432,7 +432,7 @@ u_int32_t getAddress(SFSample *sample, SFLAddress *address)
 
 char *printTag(u_int32_t tag, char *buf, int bufLen) {
   // should really be: snprintf(buf, buflen,...) but snprintf() is not always available
-  sprintf(buf, "%lu:%lu", (tag >> 12), (tag & 0x00000FFF));
+  sprintf(buf, "%lu:%lu", (unsigned long)(tag >> 12), (unsigned long)(tag & 0x00000FFF));
   return buf;
 }
